server_py/requirements.txt
flask
flask-cors
flask-jwt-extended
flask-sqlalchemy
psycopg2-binary
python-dotenv
werkzeug
